feat(usecountdown): startImmediate off by default, onDown args update  and test cases updates
startImmediate was true by default earlier which means the timer ran instantly. Now the countdown
can be configured to run immediately, or after the first interval. Our requirements generally are to
run the countdown after the first interval, but optionally this can be passed as an argument.

BREAKING CHANGE: onDown callback is only called with time remaining in milliseconds. There is no
second argument anymore as it can easily be built on the user's end with new Date()